refactor: move mutation to its own module
